requirements.txt
packaging==26.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
python-can==4.6.1
typing_extensions==4.15.0
wrapt==1.17.3
